ヘッダー部分をコンポーネント化
公式情報のお知らせ
お問い合わせ先
設定ファイルの制御(バージョン, 無効化された項目)
