+++ src/lib/components/Title.svelte
@@ -1,0 +1,1 @@
+<h2 class="mb-4 text-xl font-bold text-gray-900 dark:text-white">Palworldサーバ設定</h2>
